Still working on getting this typed.
