+++ README.md
@@ -29,3 +29,21 @@ Whenever there is a new pull request (PR) Dokploy is configured to build and dep
 
 ### Research and Ideas
 - https://www.npmjs.com/package/@lume/autolayout
+
+```javascript
+// import AutoLayout from '@lume/autolayout';
+
+// try {
+//   // The VFL can be either a string or an array of strings.
+//   // strings may also contain '\n' which indicates that a new line of VFL will begin.
+//   var constraints = AutoLayout.VisualFormat.parse([
+//     'H:|-[child1]-|',
+//     'H:|-[child2(==33%)][child3]-|',
+//     'V:|-[child1(==7.5%)][child2]-|',
+//     'V:|-[child1][child3]-|',
+//   ]);
+//   console.log(constraints);
+// } catch (err) {
+//   console.log('parse error: ' + err.toString());
+// }
+```
