+++ ui/src/pages/paper-page-components/pq-jumbotron.stories.ts
@@ -1,6 +1,6 @@
 import { html } from 'lit';
 
-import '../../../pq-jumbotron.js';
+import './pq-jumbotron.js';
 
 // More on default export: https://storybook.js.org/docs/web-components/writing-stories/introduction#default-export
 export default {
