+++ ui/src/pages/shared-components/pq-footer.stories.ts
@@ -1,6 +1,6 @@
 import { html } from 'lit';
 
-import '../../../pq-footer.js';
+import './pq-footer.js';
 
 // More on default export: https://storybook.js.org/docs/web-components/writing-stories/introduction#default-export
 export default {
