+++ ui/src/pages/shared-components/pq-menu.stories.ts
@@ -1,6 +1,6 @@
 import { html } from 'lit';
 
-import '../pq-menu.js';
+import './pq-menu.ts';
 
 // More on default export: https://storybook.js.org/docs/web-components/writing-stories/introduction#default-export
 export default {
@@ -9,6 +9,6 @@ export default {
 };
 
 // More on component templates: https://storybook.js.org/docs/web-components/writing-stories/introduction#using-args
-const Template = (args) => html`<pq-menu></pq-menu>`;
+const Template = () => html`<pq-menu></pq-menu>`;
 
 export const Default = Template.bind({});
